requirements.txt
numpy
Pillow
glfw
PyOpenGL
PyOpenGL_accelerate
PyGLM
